backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
mysql-connector-python==8.3.0
